refactor: Carousel 컴포넌트를 공통 컴포넌트에서 제외 후 피드 폴더로 이동
- API 명세의 입력값을 prop으로 받도록 변경해 common에 유지하기에는 범용성이 낮다고 판단
