+++ frontend/src/components/feed/Carousel/Carousel.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Carousel from '~/components/common/Carousel/Carousel';
+import Carousel from '~/components/feed/Carousel/Carousel';
 
 /**
  * `Carousel`는 이미지를 보여주며 자연스럽게 이미지 전환을 가능케 하는 슬라이드 쇼 컴포넌트입니다.
