+++ frontend/src/components/feed/Carousel/Carousel.styled.ts
